+++ src/init.c
@@ -68,6 +68,7 @@ extern SEXP add_();
 extern SEXP mul_();
 extern SEXP create_data_frame_in_c_();
 extern SEXP create_list_in_c_();
+extern SEXP create_array_in_c_();
 
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -79,6 +80,7 @@ static const R_CallMethodDef CEntries[] = {
   {"add_"                   , (DL_FUNC) &add_                   , 2},
   {"create_data_frame_in_c_", (DL_FUNC) &create_data_frame_in_c_, 0},
   {"create_list_in_c_"      , (DL_FUNC) &create_list_in_c_      , 0},
+  {"create_array_in_c_"     , (DL_FUNC) &create_array_in_c_     , 0},
   {NULL                     , NULL                              , 0}   // Placeholder to indicate last one.
 };